requirements.txt
mutagen
genutility
bencode.py>=2.0.0 # bencode
